refactoring file structure, add persistent base file
