+++ service/common/error_handlers.py
@@ -18,7 +18,7 @@ Module: error_handlers
 """
 from flask import jsonify
 from flask import current_app as app  # Import Flask application
-from service.models import DataValidationError
+from service.models.item import DataValidationError
 from . import status
